refactor: export getSingleViewRecursive function
